Data Analysis & AI/tests_diagnose/requirements.txt
fastapi
uvicorn
python-multipart
joblib
requests
pandas
scikit-learn
numpy
flask
python-dotenv
langchain-groq
langchain-core